# internal/domain/entities/user.go
package entities

// User represents the user model
// @Description User account information
type User struct {
	ID       string `json:"id"`
	Name     string `json:"name"`
	Email    string `json:"email"`
	Password string `json:"password"`
	Role     string `json:"role"`
}